+++ src/main/java/org/truffleruby/language/loader/FeatureLoader.java
@@ -290,6 +290,10 @@ public class FeatureLoader {
 
     @TruffleBoundary
     public List<TruffleObject> loadCExtLibrary(String feature, String path) {
+        if (!context.getOptions().CEXTS) {
+            throw new RaiseException(context, context.getCoreExceptions().loadError(String.format("cannot load %s as C extensions are disabled with -Xcexts=false", path), path, null));
+        }
+
         final File file = new File(path);
 
         if (!new File(path).exists()) {